fix auth

+++ routes/web.php
@@ -1,7 +1,6 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\UserController;
 use App\Http\Controllers\AdminController;
 use App\Http\Controllers\ClientController;
 use App\Http\Controllers\InvoiceController;
@@ -11,45 +10,21 @@ Route::get('/', function () {
     return view('auth.login');
 });
 
-
-
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
-});
-Route::middleware(['auth', 'roles:admin'])->group(function () {
-    Route::get('/admin/dashboard', [AdminController::class, 'adminDashboard'])
-         ->name('admin.dashboard');
-         //invoices
+
+    Route::get('/dashboard', [AdminController::class, 'adminDashboard'])
+        ->name('admin.dashboard');
     Route::get('/invoices', [InvoiceController::class, 'index'])->name('invoices.index');
-    Route::get('/invoices/create',[InvoiceController::class,'create'])->name('invoices.create');  
-    Route::post('/invoices/store',[InvoiceController::class,'store'])->name('invoices.store');
-    
-    Route::delete('/invoices/{id}', [InvoiceController::class, 'destroy'])->name('invoices.destroy');
-
-
-
-    
-   
+    Route::get('/invoices/create', [InvoiceController::class, 'create'])->name('invoices.create');
+    Route::get('invoices/{invoice}', [InvoiceController::class, 'show'])->name('invoices.show');
+    Route::get('/invoices/create', [InvoiceController::class, 'create'])->name('invoices.create');
+    Route::get('/invoices/{invoice}/edit', [InvoiceController::class, 'edit'])->name('invoices.edit');
+    Route::post('/invoices/store', [InvoiceController::class, 'store'])->name('invoices.store');
+    Route::put('/invoices/{invoice}', [InvoiceController::class, 'update'])->name('invoices.update');
+    Route::delete('/invoices/{id}', [InvoiceController::class, 'destroy'])->name('invoices.destroy')->middleware('isAdmin');
+    Route::resource('clients', ClientController::class);
+    Route::get('invoice/{id}/download', [InvoiceController::class, 'downloadPdf'])->name('invoice.download');
 });
-Route::get('/invoices/create',[InvoiceController::class,'create'])->name('invoices.create');  
-//clients
-Route::resource('clients', ClientController::class);
-Route::middleware(['auth','roles:user'])->group(function () {
-    Route::get('/user/dashboard', [UserController::class, 'user_dashboard'])
-         ->name('user.dashboard');
-
-
-}); 
-Route::get('invoices/{invoice}', [InvoiceController::class, 'show'])->name('invoices.show');
-Route::get('/invoices/{invoice}/edit', [InvoiceController::class, 'edit'])->name('invoices.edit');
-Route::put('/invoices/{invoice}', [InvoiceController::class, 'update'])->name('invoices.update');
-Route::get('invoice/{id}/download', [InvoiceController::class, 'downloadPdf'])->name('invoice.download');
-
-
-
-// // Include route files
-
-require __DIR__.'/auth.php';
-
